Moving verifier to common VM external interface, HARMONY-5649
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@648211 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/translator/java/JavaLabelPrepass.cpp
@@ -298,7 +298,7 @@ public:
     uint32 parseHandlers() {
         uint32 numHandlers = enclosingMethod->getNumHandlers();
         for (uint32 i=0; i<numHandlers; i++) {
-            unsigned beginOffset,endOffset,handlerOffset,handlerClassIndex;
+            unsigned short beginOffset,endOffset,handlerOffset,handlerClassIndex;
             enclosingMethod->getHandlerInfo(i,&beginOffset,&endOffset,
                 &handlerOffset,&handlerClassIndex);
             if (!catchBlock(beginOffset,endOffset-beginOffset,
